First step at two separate async monitors with the different representations. Still seriously need to consolidate those two into a better hierarchy to make it so I can just have an EnergyManager family of objects that create an energy management facility in different ways to optimize different energy services for the performance/usability/whatever else tradeoffs they provide.
Also have an async memory monitor that takes (((memory stamps))) in the background, same model as async energy monitor. created an AsyncMonitor
superclass to group and similarize the two monitors to each other and have the function as one beautiful family unit :)

Did a first stab at .ipynb files to poke around and analyze the data output of the memory experiments. Keep this around
and work on making a fresher more robust data anlysis profile, both as you get better at the data stuff, as well as the more
the experiments and incoming modes of data change and develop